+++ R/tensor.R
@@ -71,27 +71,27 @@ Tensor <- R7Class(
     },
     to = function(dtype = NULL, device = NULL, other = NULL, non_blocking = FALSE,
                   copy = FALSE, memory_format = NULL) {
-      
+
       has_device <- !is.null(device)
       has_dtype <- !is.null(dtype)
       has_other <- !is.null(other)
-      
+
       if (has_other) {
         # can't have device and dtype
         if (has_device || has_dtype) {
           cli::cli_abort("Had {.arg other} but {.arg device} or {.arg dtype} are non {.val NULL}")
         }
-        
+
         return(private$`_to`(other = other, non_blocking = non_blocking, copy = copy))
       }
-      
+
       if (!has_dtype) {
         dtype <- self$dtype
       }
-      
+
       if (has_device) {
         private$`_to`(
-          dtype = dtype, 
+          dtype = dtype,
           device = device,
           non_blocking = non_blocking,
           copy = copy,
@@ -383,7 +383,7 @@ as.matrix.torch_tensor <- function(x, ...) {
 as_array_impl <- function(x) {
   # move tensor to cpu before copying to R
   x <- x$cpu()
-  
+
   if (x$is_complex()) {
     out <- complex(
       real = as.array(x$real),
@@ -392,7 +392,7 @@ as_array_impl <- function(x) {
     dim(out) <- dim(x)
     return(out)
   }
-  
+
   a <- cpp_as_array(x)
 
   if (length(a$dim) <= 1L) {
@@ -412,10 +412,6 @@ as_array_impl <- function(x) {
 
 #' @export
 as_array.torch_tensor <- function(x) {
-  if (x$device$type == "cuda") {
-    runtime_error("Can't convert cuda tensor to R. Convert to cpu tensor before.")
-  }
-
   # dequantize before converting
   if (x$is_quantized()) {
     x <- x$dequantize()
@@ -479,13 +475,13 @@ Tensor$set("active", "ptr", function() {
 
 tensor_to_complex <- function(x) {
   torch_complex(
-    torch_tensor(Re(x), dtype = torch_double()), 
+    torch_tensor(Re(x), dtype = torch_double()),
     torch_tensor(Im(x), dtype = torch_double())
   )
 }
 
 #' Creates a tensor from a buffer of memory
-#' 
+#'
 #' It creates a tensor without taking ownership of the memory it points to.
 #' You must call `clone` if you want to copy the memory over a new tensor.
 #'