+++ R/tensor.R
@@ -33,7 +33,7 @@ Tensor <- R7Class(
       if (!is_undefined_tensor(self)) {
         if (!is.null(self$ptr$grad_fn)) {
           cat("[ grad_fn = <")
-          cat(self$ptr$grad_fn$print())
+          cat(self$ptr$grad_fn$print(newline = FALSE))
           cat("> ]")
         } else if (self$ptr$requires_grad && is.null(self$ptr$grad_fn)) {
           cat("[ requires_grad = TRUE ]")
@@ -288,6 +288,12 @@ Tensor <- R7Class(
     },
     half = function() {
       self$to(dtype=torch_half())
+    },
+    clone = function(...) {
+      x <- private$`_clone`(...)
+      attributes(x) <- attributes(self)
+
+      return(x)
     }
   ),
   active = list(