+++ NEWS.md
@@ -1,6 +1,6 @@
 # torch (development version)
 
-- Make sure deep cloning preserve state dict attributes. (#1129)
+- Make sure deep cloning preserve class attributes. (#1129)
 
 # torch 0.12.0