+++ NEWS.md
@@ -1,4 +1,4 @@
-# torch (development version)
+# torch 0.13.0
 
 ## Breaking changes